tuvi-ai/requirements.txt
lunarcalendar
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
